refactor [James]: refactor student route
